+++ src/Api/Finance/TransactionNumberGenerator/ClockTransactionNumberGenerator.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Biplane\YandexDirect\Api\Finance\TransactionNumberGenerator;
+
+use Biplane\YandexDirect\Api\Finance\TransactionNumberGenerator;
+
+use function time;
+
+final class ClockTransactionNumberGenerator implements TransactionNumberGenerator
+{
+    public function generate(): int
+    {
+        return time();
+    }
+}